update inc/imagick_install.inc fix bug
When remi repo is installed it doesn't check if remi version of imagemagick is installed and instead assumes it was installed i.e. fresh 123.09beta01 installs. However, for very old centmin mod installs prior to .07 release, remi repo wasn't installed so need to account for that.

so running centmin.sh menu option 15 will take care of it all so if remi yum repo isn't installed, install it and if remi version of imagemagick isn't installed, install it